[trunk] 'addi' fixed.

+++ func_sim/func_instr/func_instr.h
@@ -156,7 +156,7 @@ class FuncInstr
         void execute_addu()  { v_dst = v_src1 + v_src2; }
         void execute_sub()   { v_dst = (int32)v_src1 - (int32)v_src2; }
         void execute_subu()  { v_dst = v_src1 - v_src2; }
-        void execute_addi()  { v_dst = (int32)v_src1 + v_imm; }
+        void execute_addi()  { v_dst = (int32)v_src1 + (int16)v_imm; }
         void execute_addiu() { v_dst = v_src1 + v_imm; }
 
         void execute_sll()   { v_dst = v_src1 << v_imm; }
